separated operations.py/test_operations.py into multiple files. created arithmetics and constants files even though there is nothing to fill them.
