+++ heat/core/constants.py
@@ -1,0 +1,13 @@
+import itertools
+import torch
+
+from .communication import MPI
+from . import stride_tricks
+from . import types
+from . import tensor
+from .operations import __local_operation as local_op
+from .operations import __reduce_op as reduce_op
+
+__all__ = [
+
+]
