+++ heat/core/exponential.py
@@ -1,0 +1,100 @@
+import itertools
+import torch
+import warnings
+import numpy as np
+
+from .communication import MPI
+from . import stride_tricks
+from . import types
+from . import tensor
+from .operations import __local_operation as local_op
+from .operations import __reduce_op as reduce_op
+
+__all__ = [
+    'exp',
+    'log',
+    'sqrt'
+]
+
+
+def exp(x, out=None):
+    """
+    Calculate the exponential of all elements in the input array.
+
+    Parameters
+    ----------
+    x : ht.tensor
+        The value for which to compute the exponential.
+    out : ht.tensor or None, optional
+        A location in which to store the results. If provided, it must have a broadcastable shape. If not provided
+        or set to None, a fresh tensor is allocated.
+
+    Returns
+    -------
+    exponentials : ht.tensor
+        A tensor of the same shape as x, containing the positive exponentials of each element in this tensor. If out
+        was provided, logarithms is a reference to it.
+
+    Examples
+    --------
+    >>> ht.exp(ht.arange(5))
+    tensor([ 1.0000,  2.7183,  7.3891, 20.0855, 54.5981])
+    """
+    return local_op(torch.exp, x, out)
+
+
+def log(x, out=None):
+    """
+    Natural logarithm, element-wise.
+
+    The natural logarithm log is the inverse of the exponential function, so that log(exp(x)) = x. The natural
+    logarithm is logarithm in base e.
+
+    Parameters
+    ----------
+    x : ht.tensor
+        The value for which to compute the logarithm.
+    out : ht.tensor or None, optional
+        A location in which to store the results. If provided, it must have a broadcastable shape. If not provided
+        or set to None, a fresh tensor is allocated.
+
+    Returns
+    -------
+    logarithms : ht.tensor
+        A tensor of the same shape as x, containing the positive logarithms of each element in this tensor.
+        Negative input elements are returned as nan. If out was provided, logarithms is a reference to it.
+
+    Examples
+    --------
+    >>> ht.log(ht.arange(5))
+    tensor([  -inf, 0.0000, 0.6931, 1.0986, 1.3863])
+    """
+    return local_op(torch.log, x, out)
+
+
+def sqrt(x, out=None):
+    """
+    Return the non-negative square-root of a tensor element-wise.
+
+    Parameters
+    ----------
+    x : ht.tensor
+        The value for which to compute the square-roots.
+    out : ht.tensor or None, optional
+        A location in which to store the results. If provided, it must have a broadcastable shape. If not provided or
+        set to None, a fresh tensor is allocated.
+
+    Returns
+    -------
+    square_roots : ht.tensor
+        A tensor of the same shape as x, containing the positive square-root of each element in x. Negative input
+        elements are returned as nan. If out was provided, square_roots is a reference to it.
+
+    Examples
+    --------
+    >>> ht.sqrt(ht.arange(5))
+    tensor([0.0000, 1.0000, 1.4142, 1.7321, 2.0000])
+    >>> ht.sqrt(ht.arange(-5, 0))
+    tensor([nan, nan, nan, nan, nan])
+    """
+    return local_op(torch.sqrt, x, out)
