+++ heat/core/tensor.py
@@ -8,6 +8,12 @@ from . import types
 from . import devices
 from . import operations
 from . import io
+from . import constants
+from . import arithmetics
+from . import trigonometrics
+from . import exponential
+from . import rounding
+from . import reductions
 
 
 class tensor:
@@ -72,7 +78,7 @@ class tensor:
         absolute_values : ht.tensor
             A tensor containing the absolute value of each element in x.
         """
-        return operations.abs(self, out, dtype)
+        return rounding.abs(self, out, dtype)
 
     def absolute(self, out=None, dtype=None):
         """
@@ -252,6 +258,29 @@ class tensor:
         """
         return operations.copy(self)
 
+    def cos(self, out=None):
+        """
+        Return the trigonometric cosine, element-wise.
+
+        Parameters
+        ----------
+        out : ht.tensor or None, optional
+            A location in which to store the results. If provided, it must have a broadcastable shape. If not provided
+            or set to None, a fresh tensor is allocated.
+
+        Returns
+        -------
+        cosine : ht.tensor
+            A tensor of the same shape as x, containing the trigonometric cosine of each element in this tensor.
+            Negative input elements are returned as nan. If out was provided, square_roots is a reference to it.
+
+        Examples
+        --------
+        >>> ht.arange(-6, 7, 2).cos()
+        tensor([ 0.9602, -0.6536, -0.4161,  1.0000, -0.4161, -0.6536,  0.9602])
+        """
+        return trigonometrics.cos(self, out)
+
     def cpu(self):
         """
         Returns a copy of this object in main memory. If this object is already in main memory, then no copy is
@@ -362,7 +391,7 @@ class tensor:
         >>> ht.arange(5).exp()
         tensor([ 1.0000,  2.7183,  7.3891, 20.0855, 54.5981])
         """
-        return operations.exp(self, out)
+        return exponential.exp(self, out)
 
     def expand_dims(self, axis):
         # TODO: document me
@@ -403,7 +432,7 @@ class tensor:
         >>> ht.floor(ht.arange(-2.0, 2.0, 0.4))
         tensor([-2., -2., -2., -1., -1.,  0.,  0.,  0.,  1.,  1.])
         """
-        return operations.floor(self, out)
+        return rounding.floor(self, out)
 
     def log(self, out=None):
         """
@@ -429,7 +458,7 @@ class tensor:
         >>> ht.arange(5).log()
         tensor([  -inf, 0.0000, 0.6931, 1.0986, 1.3863])
         """
-        return operations.log(self, out)
+        return exponential.log(self, out)
 
     def save(self, path, *args, **kwargs):
         """
@@ -536,7 +565,7 @@ class tensor:
         >>> ht.arange(-6, 7, 2).sin()
         tensor([ 0.2794,  0.7568, -0.9093,  0.0000,  0.9093, -0.7568, -0.2794])
         """
-        return operations.sin(self, out)
+        return trigonometrics.sin(self, out)
 
     def sqrt(self, out=None):
         """
@@ -561,7 +590,7 @@ class tensor:
         >>> ht.arange(-5, 0).sqrt()
         tensor([nan, nan, nan, nan, nan])
         """
-        return operations.sqrt(self, out)
+        return exponential.sqrt(self, out)
 
     def sum(self, axis=None, out=None):
         # TODO: Allow also list of axes
@@ -596,7 +625,7 @@ class tensor:
         tensor([[[3.],
                  [3.]]])
         """
-        return operations.sum(self, axis, out)
+        return reductions.sum(self, axis, out)
 
     def transpose(self, axes=None):
         """
