+++ heat/core/tests/test_arithmetics.py
@@ -1,0 +1,10 @@
+import unittest
+import torch
+
+import heat as ht
+
+FLOAT_EPSILON = 1e-4
+
+
+class TestOperations(unittest.TestCase):
+    pass
