+++ heat/core/tests/test_operations.py
@@ -7,37 +7,6 @@ FLOAT_EPSILON = 1e-4
 
 
 class TestOperations(unittest.TestCase):
-    def test_abs(self):
-        float32_tensor = ht.arange(-10, 10, dtype=ht.float32, split=0)
-        absolute_values = ht.abs(float32_tensor)
-
-        # basic absolute test
-        self.assertIsInstance(absolute_values, ht.tensor)
-        self.assertEqual(absolute_values.dtype, ht.float32)
-        self.assertEqual(absolute_values.sum(axis=0), 100)
-
-        # check whether output works
-        output_tensor = ht.zeros(20, split=0)
-        self.assertEqual(output_tensor.sum(axis=0), 0)
-        ht.absolute(float32_tensor, out=output_tensor)
-        self.assertEqual(output_tensor.sum(axis=0), 100)
-
-        # dtype parameter
-        int64_tensor = ht.arange(-10, 10, dtype=ht.int64)
-        absolute_values = ht.abs(int64_tensor, dtype=ht.float32)
-        self.assertIsInstance(absolute_values, ht.tensor)
-        self.assertEqual(absolute_values.sum(axis=0), 100)
-        self.assertEqual(absolute_values.dtype, ht.float32)
-        self.assertEqual(absolute_values._tensor__array.dtype, torch.float32)
-
-        # exceptions
-        with self.assertRaises(TypeError):
-            ht.absolute('hello')
-        with self.assertRaises(TypeError):
-            float32_tensor.abs(out=1)
-        with self.assertRaises(TypeError):
-            float32_tensor.absolute(out=float32_tensor, dtype=3.2)
-
     def test_all(self):
         array_len = 9
 
@@ -270,124 +239,6 @@ class TestOperations(unittest.TestCase):
         with self.assertRaises(TypeError):
             ht.copy('hello world')
 
-    def test_exp(self):
-        elements = 10
-        comparison = torch.arange(elements, dtype=torch.float64).exp()
-
-        # exponential of float32
-        float32_tensor = ht.arange(elements, dtype=ht.float32)
-        float32_exp = ht.exp(float32_tensor)
-        self.assertIsInstance(float32_exp, ht.tensor)
-        self.assertEqual(float32_exp.dtype, ht.float32)
-        self.assertEqual(float32_exp.dtype, ht.float32)
-        in_range = (float32_exp._tensor__array - comparison.type(torch.float32)) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # exponential of float64
-        float64_tensor = ht.arange(elements, dtype=ht.float64)
-        float64_exp = ht.exp(float64_tensor)
-        self.assertIsInstance(float64_exp, ht.tensor)
-        self.assertEqual(float64_exp.dtype, ht.float64)
-        self.assertEqual(float64_exp.dtype, ht.float64)
-        in_range = (float64_exp._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # exponential of ints, automatic conversion to intermediate floats
-        int32_tensor = ht.arange(elements, dtype=ht.int32)
-        int32_exp = ht.exp(int32_tensor)
-        self.assertIsInstance(int32_exp, ht.tensor)
-        self.assertEqual(int32_exp.dtype, ht.float64)
-        self.assertEqual(int32_exp.dtype, ht.float64)
-        in_range = (int32_exp._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # exponential of longs, automatic conversion to intermediate floats
-        int64_tensor = ht.arange(elements, dtype=ht.int64)
-        int64_exp = ht.exp(int64_tensor)
-        self.assertIsInstance(int64_exp, ht.tensor)
-        self.assertEqual(int64_exp.dtype, ht.float64)
-        self.assertEqual(int64_exp.dtype, ht.float64)
-        in_range = (int64_exp._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.exp([1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.exp('hello world')
-
-    def test_floor(self):
-        start, end, step = -5.0, 5.0, 1.4
-        comparison = torch.arange(start, end, step, dtype=torch.float64).floor()
-
-        # exponential of float32
-        float32_tensor = ht.arange(start, end, step, dtype=ht.float32)
-        float32_floor = float32_tensor.floor()
-        self.assertIsInstance(float32_floor, ht.tensor)
-        self.assertEqual(float32_floor.dtype, ht.float32)
-        self.assertEqual(float32_floor.dtype, ht.float32)
-        self.assertTrue((float32_floor._tensor__array == comparison.type(torch.float32)).all())
-
-        # exponential of float64
-        float64_tensor = ht.arange(start, end, step, dtype=ht.float64)
-        float64_floor = float64_tensor.floor()
-        self.assertIsInstance(float64_floor, ht.tensor)
-        self.assertEqual(float64_floor.dtype, ht.float64)
-        self.assertEqual(float64_floor.dtype, ht.float64)
-        self.assertTrue((float64_floor._tensor__array == comparison).all())
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.floor([0, 1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.floor(object())
-
-    def test_log(self):
-        elements = 15
-        comparison = torch.arange(1, elements, dtype=torch.float64).log()
-
-        # logarithm of float32
-        float32_tensor = ht.arange(1, elements, dtype=ht.float32)
-        float32_log = ht.log(float32_tensor)
-        self.assertIsInstance(float32_log, ht.tensor)
-        self.assertEqual(float32_log.dtype, ht.float32)
-        self.assertEqual(float32_log.dtype, ht.float32)
-        in_range = (float32_log._tensor__array - comparison.type(torch.float32)) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # logarithm of float64
-        float64_tensor = ht.arange(1, elements, dtype=ht.float64)
-        float64_log = ht.log(float64_tensor)
-        self.assertIsInstance(float64_log, ht.tensor)
-        self.assertEqual(float64_log.dtype, ht.float64)
-        self.assertEqual(float64_log.dtype, ht.float64)
-        in_range = (float64_log._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # logarithm of ints, automatic conversion to intermediate floats
-        int32_tensor = ht.arange(1, elements, dtype=ht.int32)
-        int32_log = ht.log(int32_tensor)
-        self.assertIsInstance(int32_log, ht.tensor)
-        self.assertEqual(int32_log.dtype, ht.float64)
-        self.assertEqual(int32_log.dtype, ht.float64)
-        in_range = (int32_log._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # logarithm of longs, automatic conversion to intermediate floats
-        int64_tensor = ht.arange(1, elements, dtype=ht.int64)
-        int64_log = ht.log(int64_tensor)
-        self.assertIsInstance(int64_log, ht.tensor)
-        self.assertEqual(int64_log.dtype, ht.float64)
-        self.assertEqual(int64_log.dtype, ht.float64)
-        in_range = (int64_log._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.log([1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.log('hello world')
-
     def test_max(self):
         data = [
             [1,   2,  3],
@@ -540,247 +391,6 @@ class TestOperations(unittest.TestCase):
         with self.assertRaises(ValueError):
             ht.min(ht_array, axis=-4)
 
-    def test_sin(self):
-        # base elements
-        elements = 30
-        comparison = torch.arange(elements, dtype=torch.float64).sin()
-
-        # sine of float32
-        float32_tensor = ht.arange(elements, dtype=ht.float32)
-        float32_sin = ht.sin(float32_tensor)
-        self.assertIsInstance(float32_sin, ht.tensor)
-        self.assertEqual(float32_sin.dtype, ht.float32)
-        self.assertEqual(float32_sin.dtype, ht.float32)
-        in_range = (float32_sin._tensor__array - comparison.type(torch.float32)) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # sine of float64
-        float64_tensor = ht.arange(elements, dtype=ht.float64)
-        float64_sin = ht.sin(float64_tensor)
-        self.assertIsInstance(float64_sin, ht.tensor)
-        self.assertEqual(float64_sin.dtype, ht.float64)
-        self.assertEqual(float64_sin.dtype, ht.float64)
-        in_range = (float64_sin._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # logarithm of ints, automatic conversion to intermediate floats
-        int32_tensor = ht.arange(elements, dtype=ht.int32)
-        int32_sin = ht.sin(int32_tensor)
-        self.assertIsInstance(int32_sin, ht.tensor)
-        self.assertEqual(int32_sin.dtype, ht.float64)
-        self.assertEqual(int32_sin.dtype, ht.float64)
-        in_range = (int32_sin._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # logathm of longs, automatic conversion to intermediate floats
-        int64_tensor = ht.arange(elements, dtype=ht.int64)
-        int64_sin = ht.sin(int64_tensor)
-        self.assertIsInstance(int64_sin, ht.tensor)
-        self.assertEqual(int64_sin.dtype, ht.float64)
-        self.assertEqual(int64_sin.dtype, ht.float64)
-        in_range = (int64_sin._tensor__array - comparison) < FLOAT_EPSILON
-        self.assertTrue(in_range.all())
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.sin([1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.sin('hello world')
-
-    def test_sqrt(self):
-        elements = 25
-        comparison = torch.arange(elements, dtype=torch.float64).sqrt()
-
-        # square roots of float32
-        float32_tensor = ht.arange(elements, dtype=ht.float32)
-        float32_sqrt = ht.sqrt(float32_tensor)
-        self.assertIsInstance(float32_sqrt, ht.tensor)
-        self.assertEqual(float32_sqrt.dtype, ht.float32)
-        self.assertEqual(float32_sqrt.dtype, ht.float32)
-        self.assertTrue(torch.allclose(float32_sqrt._tensor__array, comparison.type(torch.float32), 1e-06))
-
-        # square roots of float64
-        float64_tensor = ht.arange(elements, dtype=ht.float64)
-        float64_sqrt = ht.sqrt(float64_tensor)
-        self.assertIsInstance(float64_sqrt, ht.tensor)
-        self.assertEqual(float64_sqrt.dtype, ht.float64)
-        self.assertEqual(float64_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(float64_sqrt._tensor__array, comparison, 1e-06))
-
-        # square roots of ints, automatic conversion to intermediate floats
-        int32_tensor = ht.arange(elements, dtype=ht.int32)
-        int32_sqrt = ht.sqrt(int32_tensor)
-        self.assertIsInstance(int32_sqrt, ht.tensor)
-        self.assertEqual(int32_sqrt.dtype, ht.float64)
-        self.assertEqual(int32_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(int32_sqrt._tensor__array, comparison, 1e-06))
-
-        # square roots of longs, automatic conversion to intermediate floats
-        int64_tensor = ht.arange(elements, dtype=ht.int64)
-        int64_sqrt = ht.sqrt(int64_tensor)
-        self.assertIsInstance(int64_sqrt, ht.tensor)
-        self.assertEqual(int64_sqrt.dtype, ht.float64)
-        self.assertEqual(int64_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(int64_sqrt._tensor__array, comparison, 1e-06))
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.sqrt([1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.sqrt('hello world')
-
-    def test_sqrt_method(self):
-        elements = 25
-        comparison = torch.arange(elements, dtype=torch.float64).sqrt()
-
-        # square roots of float32
-        float32_sqrt = ht.arange(elements, dtype=ht.float32).sqrt()
-        self.assertIsInstance(float32_sqrt, ht.tensor)
-        self.assertEqual(float32_sqrt.dtype, ht.float32)
-        self.assertEqual(float32_sqrt.dtype, ht.float32)
-        self.assertTrue(torch.allclose(float32_sqrt._tensor__array, comparison.type(torch.float32), 1e-05))
-
-        # square roots of float64
-        float64_sqrt = ht.arange(elements, dtype=ht.float64).sqrt()
-        self.assertIsInstance(float64_sqrt, ht.tensor)
-        self.assertEqual(float64_sqrt.dtype, ht.float64)
-        self.assertEqual(float64_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(float64_sqrt._tensor__array, comparison, 1e-05))
-
-        # square roots of ints, automatic conversion to intermediate floats
-        int32_sqrt = ht.arange(elements, dtype=ht.int32).sqrt()
-        self.assertIsInstance(int32_sqrt, ht.tensor)
-        self.assertEqual(int32_sqrt.dtype, ht.float64)
-        self.assertEqual(int32_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(int32_sqrt._tensor__array, comparison, 1e-05))
-
-        # square roots of longs, automatic conversion to intermediate floats
-        int64_sqrt = ht.arange(elements, dtype=ht.int64).sqrt()
-        self.assertIsInstance(int64_sqrt, ht.tensor)
-        self.assertEqual(int64_sqrt.dtype, ht.float64)
-        self.assertEqual(int64_sqrt.dtype, ht.float64)
-        self.assertTrue(torch.allclose(int64_sqrt._tensor__array, comparison, 1e-05))
-
-        # check exceptions
-        with self.assertRaises(TypeError):
-            ht.sqrt([1, 2, 3])
-        with self.assertRaises(TypeError):
-            ht.sqrt('hello world')
-
-    def test_sqrt_out_of_place(self):
-        elements = 30
-        output_shape = (3, elements)
-        number_range = ht.arange(elements, dtype=ht.float32)
-        output_buffer = ht.zeros(output_shape, dtype=ht.float32)
-
-        # square roots
-        float32_sqrt = ht.sqrt(number_range, out=output_buffer)
-        comparison = torch.arange(elements, dtype=torch.float32).sqrt()
-
-        # check whether the input range remain unchanged
-        self.assertIsInstance(number_range, ht.tensor)
-        self.assertEqual(number_range.sum(axis=0), 190)  # gaussian sum
-        self.assertEqual(number_range.gshape, (elements,))
-
-        # check whether the output buffer still has the correct shape
-        self.assertIsInstance(float32_sqrt, ht.tensor)
-        self.assertEqual(float32_sqrt.dtype, ht.float32)
-        self.assertEqual(float32_sqrt._tensor__array.shape, output_shape)
-        for row in range(output_shape[0]):
-            self.assertTrue((float32_sqrt._tensor__array[row] == comparison).all())
-
-        # exception
-        with self.assertRaises(TypeError):
-            ht.sqrt(number_range, 'hello world')
-
-    def test_sum(self):
-        array_len = 11
-
-        # check sum over all float elements of 1d tensor locally 
-        shape_noaxis = ht.ones(array_len)
-        no_axis_sum = shape_noaxis.sum()
-
-        self.assertIsInstance(no_axis_sum, ht.tensor)
-        self.assertEqual(no_axis_sum.shape, (1,))
-        self.assertEqual(no_axis_sum.lshape, (1,))
-        self.assertEqual(no_axis_sum.dtype, ht.float32)
-        self.assertEqual(no_axis_sum._tensor__array.dtype, torch.float32)
-        self.assertEqual(no_axis_sum.split, None)
-        self.assertEqual(no_axis_sum._tensor__array, array_len)
-
-        out_noaxis = ht.zeros((1,))
-        ht.sum(shape_noaxis, out=out_noaxis)
-        self.assertTrue(out_noaxis._tensor__array == shape_noaxis._tensor__array.sum())
-
-        # check sum over all float elements of split 1d tensor
-        shape_noaxis_split = ht.arange(array_len, split=0)
-        shape_noaxis_split_sum = shape_noaxis_split.sum()
-
-        self.assertIsInstance(shape_noaxis_split_sum, ht.tensor)
-        self.assertEqual(shape_noaxis_split_sum.shape, (1,))
-        self.assertEqual(shape_noaxis_split_sum.lshape, (1,))
-        self.assertEqual(shape_noaxis_split_sum.dtype, ht.int64)
-        self.assertEqual(shape_noaxis_split_sum._tensor__array.dtype, torch.int64)
-        self.assertEqual(shape_noaxis_split_sum.split, None)
-        self.assertEqual(shape_noaxis_split_sum, 55)
-
-        out_noaxis = ht.zeros((1,))
-        ht.sum(shape_noaxis_split, out=out_noaxis)
-        self.assertEqual(out_noaxis._tensor__array, 55)
-
-        # check sum over all float elements of 3d tensor locally
-        shape_noaxis = ht.ones((3, 3, 3))
-        no_axis_sum = shape_noaxis.sum()
-
-        self.assertIsInstance(no_axis_sum, ht.tensor)
-        self.assertEqual(no_axis_sum.shape, (1,))
-        self.assertEqual(no_axis_sum.lshape, (1,))
-        self.assertEqual(no_axis_sum.dtype, ht.float32)
-        self.assertEqual(no_axis_sum._tensor__array.dtype, torch.float32)
-        self.assertEqual(no_axis_sum.split, None)
-        self.assertEqual(no_axis_sum._tensor__array, 27)
-
-        out_noaxis = ht.zeros((1,))
-        ht.sum(shape_noaxis, out=out_noaxis)
-        self.assertEqual(out_noaxis._tensor__array, 27)
-
-        # check sum over all float elements of split 3d tensor
-        shape_noaxis_split_axis = ht.ones((3, 3, 3), split=0)
-        split_axis_sum = shape_noaxis_split_axis.sum(axis=0)
-
-        self.assertIsInstance(split_axis_sum, ht.tensor)
-        self.assertEqual(split_axis_sum.shape, (1, 3, 3))
-        self.assertEqual(split_axis_sum.dtype, ht.float32)
-        self.assertEqual(split_axis_sum._tensor__array.dtype, torch.float32)
-        self.assertEqual(split_axis_sum.split, None)
-
-        out_noaxis = ht.zeros((1, 3, 3,))
-        ht.sum(shape_noaxis, axis=0, out=out_noaxis)
-        self.assertTrue((out_noaxis._tensor__array == torch.full((1, 3, 3,), 3)).all())
-
-        # check sum over all float elements of splitted 5d tensor with negative axis
-        shape_noaxis_split_axis_neg = ht.ones((1, 2, 3, 4, 5), split=1)
-        shape_noaxis_split_axis_neg_sum = shape_noaxis_split_axis_neg.sum(axis=-2)
-
-        self.assertIsInstance(shape_noaxis_split_axis_neg_sum, ht.tensor)
-        self.assertEqual(shape_noaxis_split_axis_neg_sum.shape, (1, 2, 3, 1, 5))
-        self.assertEqual(shape_noaxis_split_axis_neg_sum.dtype, ht.float32)
-        self.assertEqual(shape_noaxis_split_axis_neg_sum._tensor__array.dtype, torch.float32)
-        self.assertEqual(shape_noaxis_split_axis_neg_sum.split, 1)
-
-        out_noaxis = ht.zeros((1, 2, 3, 1, 5))
-        ht.sum(shape_noaxis_split_axis_neg, axis=-2, out=out_noaxis)
-
-        # exceptions
-        with self.assertRaises(ValueError):
-            ht.ones(array_len).sum(axis=1)
-        with self.assertRaises(ValueError):
-            ht.ones(array_len).sum(axis=-2)
-        with self.assertRaises(ValueError):
-            ht.ones((4, 4)).sum(axis=0, out=out_noaxis)
-        with self.assertRaises(TypeError):
-            ht.ones(array_len).sum(axis='bad_axis_type')
-            
     def test_transpose(self):
         # vector transpose, not distributed
         vector = ht.arange(10)
