+++ heat/core/trigonometrics.py
@@ -1,0 +1,68 @@
+import itertools
+import torch
+import warnings
+import numpy as np
+
+from .communication import MPI
+from . import stride_tricks
+from . import types
+from . import tensor
+from .operations import __local_operation as local_op
+from .operations import __reduce_op as reduce_op
+
+__all__ = [
+    'sin',
+    'cos'
+]
+
+
+def sin(x, out=None):
+    """
+    Return the trigonometric sine, element-wise.
+
+    Parameters
+    ----------
+    x : ht.tensor
+        The value for which to compute the trigonometric sine.
+    out : ht.tensor or None, optional
+        A location in which to store the results. If provided, it must have a broadcastable shape. If not provided
+        or set to None, a fresh tensor is allocated.
+
+    Returns
+    -------
+    sine : ht.tensor
+        A tensor of the same shape as x, containing the trigonometric sine of each element in this tensor.
+        Negative input elements are returned as nan. If out was provided, square_roots is a reference to it.
+
+    Examples
+    --------
+    >>> ht.sin(ht.arange(-6, 7, 2))
+    tensor([ 0.2794,  0.7568, -0.9093,  0.0000,  0.9093, -0.7568, -0.2794])
+    """
+    return local_op(torch.sin, x, out)
+
+
+def cos(x, out=None):
+    """
+    Return the trigonometric cosine, element-wise.
+
+    Parameters
+    ----------
+    x : ht.tensor
+        The value for which to compute the trigonometric sine.
+    out : ht.tensor or None, optional
+        A location in which to store the results. If provided, it must have a broadcastable shape. If not provided
+        or set to None, a fresh tensor is allocated.
+
+    Returns
+    -------
+    cosine : ht.tensor
+        A tensor of the same shape as x, containing the trigonometric cosine of each element in this tensor.
+        Negative input elements are returned as nan. If out was provided, square_roots is a reference to it.
+
+    Examples
+    --------
+    >>> ht.cos(ht.arange(-6, 7, 2))
+    tensor([ 0.2794,  0.7568, -0.9093,  0.0000,  0.9093, -0.7568, -0.2794])
+    """
+    return local_op(torch.cos, x, out)
